django-fyp-new-30-April/requirements.txt
python-dotenv
google-generativeai
langchain
langchain-google-genai
langchain-community
faiss-cpu
transformers
django
pypdf
sentence-transformers